# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Epic Poetry Cafe Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settlements:
    get:
      operationId: listSettlements
      tags: [settlements]
      summary: List settlements
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
        - name: toDate
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Settlements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Settlement"
    post:
      operationId: createSettlement
      tags: [settlements]
      summary: Create settlement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSettlementRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
  /settlements/{id}:
    get:
      operationId: getSettlement
      tags: [settlements]
      summary: Get settlement with lines
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Settlement detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementDetail"
    patch:
      operationId: updateSettlement
      tags: [settlements]
      summary: Update settlement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSettlementRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
    delete:
      operationId: deleteSettlement
      tags: [settlements]
      summary: Delete settlement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /settlements/{id}/verify:
    post:
      operationId: verifySettlement
      tags: [settlements]
      summary: Verify settlement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
  /settlements/sales-summary:
    get:
      operationId: getSettlementSalesSummary
      tags: [settlements]
      summary: Get sales summary for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementSalesSummary"
...
components:
  schemas:
...
    Settlement:
      type: object
      properties:
        id:
          type: integer
        settlementDate:
          type: string
        grossSalesAmount:
          type: number
        discountAmount:
          type: number
        netSalesAmount:
          type: number
        totalSettlementAmount:
          type: number
        differenceAmount:
          type: number
        differenceType:
          type: string
        status:
          type: string
        remarks:
          type: ["string", "null"]
        createdBy:
          type: ["integer", "null"]
        verifiedBy:
          type: ["integer", "null"]
        verifiedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - settlementDate
        - grossSalesAmount
        - netSalesAmount
        - totalSettlementAmount
        - differenceAmount
        - differenceType
        - status
    SettlementLine:
      type: object
      properties:
        id:
          type: integer
        settlementId:
          type: integer
        paymentMode:
          type: string
        amount:
          type: number
        referenceNote:
          type: ["string", "null"]
      required:
        - id
        - settlementId
        - paymentMode
        - amount
    SettlementDetail:
      type: object
      properties:
        settlement:
          $ref: "#/components/schemas/Settlement"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/SettlementLine"
      required:
        - settlement
        - lines
    CreateSettlementRequest:
      type: object
      properties:
        settlementDate:
          type: string
        remarks:
          type: string
        lines:
          type: array
          items:
            type: object
            properties:
              paymentMode:
                type: string
              amount:
                type: number
              referenceNote:
                type: string
            required:
              - paymentMode
              - amount
      required:
        - settlementDate
        - lines
    SettlementSalesSummary:
      type: object
      properties:
        date:
          type: string
        grossSales:
          type: number
        totalDiscount:
          type: number
        netSales:
          type: number
        itemCount:
          type: integer
      required:
        - date
        - grossSales
        - totalDiscount
        - netSales
        - itemCount
...