# lib/api-zod/src/generated/types/createSettlementRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Epic Poetry Cafe Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSettlementRequestLinesItem } from "./createSettlementRequestLinesItem";

export interface CreateSettlementRequest {
  settlementDate: string;
  remarks?: string;
  lines: CreateSettlementRequestLinesItem[];
}