# lib/api-zod/src/generated/types/settlementDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Epic Poetry Cafe Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Settlement } from "./settlement";
import type { SettlementLine } from "./settlementLine";

export interface SettlementDetail {
  settlement: Settlement;
  lines: SettlementLine[];
}