# lib/api-zod/src/generated/types/settlementLine.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Epic Poetry Cafe Management API
 * OpenAPI spec version: 0.1.0
 */

export interface SettlementLine {
  id: number;
  settlementId: number;
  paymentMode: string;
  amount: number;
  /** @nullable */
  referenceNote?: string | null;
}