# lib/api-zod/src/generated/types/settlementSalesSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Epic Poetry Cafe Management API
 * OpenAPI spec version: 0.1.0
 */

export interface SettlementSalesSummary {
  date: string;
  grossSales: number;
  totalDiscount: number;
  netSales: number;
  itemCount: number;
}